+++ AfxHookSource2/MirvPovVoice.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <Windows.h>
+
+void MirvPov_HookVoiceHud(HMODULE clientDll);
+void MirvPov_ResetVoiceHud();
+void MirvPov_UpdateVoiceTeam();
+void MirvPov_UpdateVoiceHud();
+void MirvPov_ClearSyntheticSpeaking();
